pyproject.toml
[project]
name = "rhodo"
version = "0.1.0"
description = "Discord bot"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "aiohttp",
  "arkprts[assets]",
  "beautifulsoup4",
  "brotli",
  "chromadb",
  "discord-py==2.7.1",
  "flask",
  "langchain",
  "langchain-chroma",
  "langchain-community",
  "langchain-huggingface",
  "langchain-openai",
  "langchain-core>=0.3.72",
  "langgraph",
  "loguru",
  "lxml",
  "onnxocr",
  "opencv-python",
  "Pillow",
  "PyNaCl",
  "python-dotenv",
  "requests",
  "tavily-python",
  "tweety-ns",
  "waitress",
  "zstandard",
  "sentence-transformers>=5.2.0",
  "sentencepiece>=0.2.1",
]
